# go/docs/swagger.yaml
basePath: /
definitions:
...
  enums.SubmissionStatus:
    enum:
    - pending
    - running
    - error
    - completed
    type: string
    x-enum-varnames:
    - SubmissionPending
    - SubmissionRunning
    - SubmissionError
    - SubmissionCompleted
...
  models.Submission:
    properties:
      code:
        description: สำหรับ code exercises
        type: string
      error_message:
        type: string
      failed_count:
        type: integer
      file_name:
        description: ชื่อไฟล์ที่ส่ง
        type: string
      file_size:
        description: ขนาดไฟล์
        type: integer
      file_url:
        description: สำหรับ PDF exercises
        type: string
      material_id:
        description: สำหรับ course materials
        type: string
      mime_type:
        description: ประเภทไฟล์
        type: string
      passed_count:
        type: integer
      results:
        items:
          $ref: '#/definitions/models.SubmissionResult'
        type: array
      status:
        $ref: '#/definitions/enums.SubmissionStatus'
      submission_id:
        type: string
      submitted_at:
        type: string
      total_score:
        type: integer
      user_id:
        type: string
    type: object
  models.SubmissionResult:
    properties:
      actual_output:
        items:
          type: integer
        type: array
      createdAt:
        type: string
      error_message:
        type: string
      result_id:
        type: string
      status:
        type: string
      submission_id:
        type: string
      test_case_id:
        type: string
    type: object
...
  response.StandardResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: 65070209@kmitl.ac.th
    name: API Support
    url: https://github.com/Project-DSView/backend
  description: This is the backend API for DSView application In Services GO
  termsOfService: http://swagger.io/terms/
  title: DSView API Go Service
  version: 1.0.0-alpha
paths:
...
  /api/exercises/{id}/submissions:
    get:
      description: ดูการส่งงานของแบบฝึกหัด (Teachers และ TAs ที่ enroll ในคอร์สเท่านั้น)
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  pagination:
                    properties:
                      limit:
                        type: integer
                      page:
                        type: integer
                      total:
                        type: integer
                      total_pages:
                        type: integer
                    type: object
                  submissions:
                    items:
                      properties:
                        code:
                          type: string
                        failedCount:
                          type: integer
                        passedCount:
                          type: integer
                        status:
                          type: string
                        studentName:
                          type: string
                        submissionId:
                          type: string
                        submittedAt:
                          type: string
                        totalScore:
                          type: integer
                        userId:
                          type: string
                      type: object
                    type: array
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: List exercise submissions
      tags:
      - submissions
  /api/exercises/{id}/submit:
    post:
      consumes:
      - application/json
      description: ส่งแบบฝึกหัด Python และประเมินผลด้วย Docker executor เทียบกับ test
        cases
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                properties:
                  code:
                    type: string
                  exerciseId:
                    type: string
                  failedCount:
                    type: integer
                  passedCount:
                    type: integer
                  results:
                    items:
                      properties:
                        actualOutput:
                          type: object
                        errorMessage:
                          type: string
                        resultId:
                          type: string
                        status:
                          type: string
                        testCaseId:
                          type: string
                      type: object
                    type: array
                  status:
                    type: string
                  submissionId:
                    type: string
                  submittedAt:
                    type: string
                  totalScore:
                    type: integer
                  userId:
                    type: string
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Submit exercise
      tags:
      - submissions
...
  /api/submissions/{id}:
    get:
      description: ดูรายละเอียดการส่งงาน (เจ้าของ, Teachers, หรือ TAs ที่ enroll ในคอร์สเท่านั้น)
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  code:
                    type: string
                  errorMessage:
                    type: string
                  exerciseId:
                    type: string
                  failedCount:
                    type: integer
                  passedCount:
                    type: integer
                  results:
                    items:
                      properties:
                        actualOutput:
                          type: object
                        errorMessage:
                          type: string
                        resultId:
                          type: string
                        status:
                          type: string
                        testCaseId:
                          type: string
                      type: object
                    type: array
                  status:
                    type: string
                  submissionId:
                    type: string
                  submittedAt:
                    type: string
                  totalScore:
                    type: integer
                  userId:
                    type: string
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Submission not found
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Get submission detail
      tags:
      - submissions
  /api/submissions/{submission_id}:
    get:
      description: Get a specific PDF submission
      parameters:
      - description: Submission ID
        in: path
        name: submission_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Submission'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get PDF submission
      tags:
      - pdf-exercises
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication for secure endpoints.
    in: header
    name: dsview-api-key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"